Refactor into audio, lights, and oracle modules
